+++ examples/05-window-manager-shell/src/index.ts
@@ -1,0 +1,1 @@
+export { WindowManagerShellExample } from './WindowManagerShellExample';